Fix address/bias on 32-bit systems
More generally, any platform where library addresses can be in the
"negative" part of the address space (ie, address MSB is 1).

This consistently uses `usize` for both `Bias` and
addresses (`Avma`, `Svma`). This assumes `Bias` is always positive, and
can be very large.

Converting `Avma`/`Svma` to usize makes the address
arithmetic easier, as it avoids `ptr::offset`, which can take both
positive and negative offsets, which means it can't offset by more than
half the address space. It is also undefined for general `*mut u8`
pointer arithmetic ("Both the starting and resulting pointer must
be either in bounds or one byte past the end of the same allocated
object" is not met, nor is "The offset being in bounds cannot rely on
"wrapping around" the address space. That is, the infinite-precision
sum, in bytes must fit in a usize").

Fixes issue #47