feat: improve handling of pdb filename from codeview
Do not use a slice on the CodeViewRecord70 struct for the pdb_filename,
this doesn't really match the underlying storage. This a flexible array
storing a c-str, which we cannot really represent in Rust.
Instead, only represent the first byte, and parse the c-str from its
address when needed.
This is still unsafe though, as the CodeViewRecord70 could be
copied/moved. A new CodeView struct could be created, or the one
returned could be pinned.